Clean GetPointInteractorの冗長なit.toInt()呼び出しを削除
getNumberOfTicket()がFlow<Int>を返すため、map内のitは既にInt型であり
toInt()は冗長。NumberOfTicketへ直接渡すよう簡略化する。

+++ application/reward/src/main/java/jp/kztproject/rewardedtodo/application/reward/usecase/GetPointInteractor.kt
@@ -8,6 +8,6 @@ import javax.inject.Inject
 
 class GetPointInteractor @Inject constructor(private val ticketRepository: ITicketRepository) : GetPointUseCase {
     override suspend fun execute(): Flow<NumberOfTicket> = ticketRepository.getNumberOfTicket().map {
-        NumberOfTicket(it.toInt())
+        NumberOfTicket(it)
     }
 }
